CustomJob class: remove other implementations.
